Replace rand with fastrand in simulation

+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const PLAYER_CARDS: i8 = 2;
-pub const COMMON_CARDS: i8 = 5;
+pub const PLAYER_CARDS: usize = 2;
+pub const COMMON_CARDS: usize = 5;
